backend/requirements.txt
Django==1.11.15
Django==2.1
Django==4.1.3
djangorestframework==3.14.0
numpy==1.15.1
numpy==1.20.3
opencv_python==4.5.5.62
Pillow==8.4.0
Pillow==10.0.0
selenium==4.10.0
torch==1.11.0
torchvision==0.12.0
